# Conflicts:
#	src/android_app/app/src/main/AndroidManifest.xml
#	src/android_app/app/src/main/java/com/mobilehealthsports/vaccinepass/ui/main/add_vaccine/AddVaccineFragment.kt

+++ src/android_app/app/src/main/java/com/mobilehealthsports/vaccinepass/business/models/Vaccine.kt
@@ -4,7 +4,7 @@ import androidx.room.PrimaryKey
 import java.util.*
 
 data class Vaccine(
-    val uid: Int,
+    val uid: Long,
     val name: String,
     val company: String?,
     val indication: String?,
